type validation wip
