+++ include/bx/engine/modules/graphics/backend/graphics_opengl.hpp
@@ -2,33 +2,35 @@
 
 #include "bx/engine/modules/graphics.hpp"
 
-#include "bx/engine/containers/string.hpp"
-#include "bx/engine/containers/hash_map.hpp"
+#include "opengl/opengl_api.hpp"
 
-#include <glad/glad.h>
-
-#ifdef BX_GRAPHICS_OPENGL_BACKEND
-#define GLSL_VERSION "#version 460 core\n"
-#endif
-
-#ifdef BX_GRAPHICS_OPENGLES_BACKEND
-#define GLSL_VERSION "#version 310 es\n"
-#endif
-
-#define GLSL_VERT_SHADER GLSL_VERSION "#define VERTEX\n"
-#define GLSL_FRAG_SHADER GLSL_VERSION "#define PIXEL\n"
-
-#define GPU_MEMORY_INFO_DEDICATED_VIDMEM_NVX            0x9047
-#define GPU_MEMORY_INFO_TOTAL_AVAILABLE_MEMORY_NVX      0x9048
-#define GPU_MEMORY_INFO_CURRENT_AVAILABLE_VIDMEM_NVX    0x9049
-#define GPU_MEMORY_INFO_EVICTION_COUNT_NVX              0x904A
-#define GPU_MEMORY_INFO_EVICTED_MEMORY_NVX              0x904B
-
-#define VBO_FREE_MEMORY_ATI                             0x87FB
-#define TEXTURE_FREE_MEMORY_ATI                         0x87FC
-#define RENDERBUFFER_FREE_MEMORY_ATI                    0x87FD
-
-#define MAX_BOUND_VERTEX_BUFFERS                        16
+//#include "bx/engine/containers/string.hpp"
+//#include "bx/engine/containers/hash_map.hpp"
+//
+//#include <glad/glad.h>
+//
+//#ifdef BX_GRAPHICS_OPENGL_BACKEND
+//#define GLSL_VERSION "#version 460 core\n"
+//#endif
+//
+//#ifdef BX_GRAPHICS_OPENGLES_BACKEND
+//#define GLSL_VERSION "#version 310 es\n"
+//#endif
+//
+//#define GLSL_VERT_SHADER GLSL_VERSION "#define VERTEX\n"
+//#define GLSL_FRAG_SHADER GLSL_VERSION "#define PIXEL\n"
+//
+//#define GPU_MEMORY_INFO_DEDICATED_VIDMEM_NVX            0x9047
+//#define GPU_MEMORY_INFO_TOTAL_AVAILABLE_MEMORY_NVX      0x9048
+//#define GPU_MEMORY_INFO_CURRENT_AVAILABLE_VIDMEM_NVX    0x9049
+//#define GPU_MEMORY_INFO_EVICTION_COUNT_NVX              0x904A
+//#define GPU_MEMORY_INFO_EVICTED_MEMORY_NVX              0x904B
+//
+//#define VBO_FREE_MEMORY_ATI                             0x87FB
+//#define TEXTURE_FREE_MEMORY_ATI                         0x87FC
+//#define RENDERBUFFER_FREE_MEMORY_ATI                    0x87FD
+//
+//#define MAX_BOUND_VERTEX_BUFFERS                        16
 
 //struct ShaderImpl
 //{
