+++ include/bx/engine/modules/graphics/backend/opengl/conversion.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "bx/engine/core/guard.hpp"
+#include "bx/engine/containers/string.hpp"
+
+#include "opengl_api.hpp"
+#include "bx/engine/modules/graphics/type.hpp"
+
+namespace Gl
+{
+}
