+++ src/bx/engine/modules/graphics/backend/opengl/conversion.cpp
@@ -1,0 +1,7 @@
+#include "bx/engine/modules/graphics/backend/opengl/conversion.hpp"
+
+#include "bx/engine/core/macros.hpp"
+
+namespace Gl
+{
+}
