+++ src/bx/engine/modules/graphics/type.cpp
@@ -1,14 +1,14 @@
 #include "bx/engine/modules/graphics/type.hpp"
 
-u64 HBuffer::null = { 0 };
-u64 HTexture::null = { 0 };
-u64 HTextureView::null = { 0 };
-u64 HShader::null = { 0 };
-u64 HGraphicsPipeline::null = { 0 };
-u64 HComputePipeline::null = { 0 };
-u64 HRenderPass::null = { 0 };
-u64 HBindGroupLayout::null = { 0 };
-u64 HBindGroup::null = { 0 };
+const HBuffer HBuffer::null = { 0 };
+const HTexture HTexture::null = { 0 };
+const HTextureView HTextureView::null = { 0 };
+const HShader HShader::null = { 0 };
+const HGraphicsPipeline HGraphicsPipeline::null = { 0 };
+const HComputePipeline HComputePipeline::null = { 0 };
+const HRenderPass HRenderPass::null = { 0 };
+const HBindGroupLayout HBindGroupLayout::null = { 0 };
+const HBindGroup HBindGroup::null = { 0 };
 
 BlendComponent BlendComponent::Replace()
 {
